+++ src/tools/voxedit/modules/voxedit-ui/LSystemPanel.h
@@ -4,30 +4,23 @@
 
 #pragma once
 
+#include "core/collection/DynamicArray.h"
 #include "ui/Panel.h"
-#include "core/String.h"
 #include "voxedit-util/SceneManager.h"
+#include "voxelgenerator/LSystem.h"
 
 namespace voxedit {
 
 class LSystemPanel : public ui::Panel {
 private:
 	using Super = ui ::Panel;
-	struct LSystemData {
-		core::String axiom = "F";
-		core::String rulesStr = R"({
-	F
-	(67)F+[!+F-F-F(37)L]-[!-F+F+F(142)L]>[!F<F<F(128)L]<[!<F>F>F(123)L]
-})";
-		float angle = 22.5f;
-		float length = 12.0f;
-		float width = 4.0f;
-		float widthIncrement = 1.5f;
-		int iterations = 2;
-		float leavesRadius = 8.0f;
-	};
-	LSystemData _lsystemData;
+	core::DynamicArray<voxelgenerator::lsystem::LSystemTemplate> _templates;
+	voxelgenerator::lsystem::LSystemConfig _conf;
 	SceneManagerPtr _sceneMgr;
+	int _templateIdx = -1;
+
+	void copyRulesToClipboard();
+	void pasteRulesFromClipboard();
 
 public:
 	LSystemPanel(ui::IMGUIApp *app, const SceneManagerPtr &sceneMgr) : Super(app, "lsystem"), _sceneMgr(sceneMgr) {
@@ -40,4 +33,4 @@ public:
 #endif
 };
 
-}
+} // namespace voxedit