+++ src/tools/voxedit/modules/voxedit-ui/MenuBar.cpp
@@ -111,7 +111,7 @@ bool MenuBar::update(ui::IMGUIApp *app, command::CommandExecutionListener &liste
 			ImGui::CommandIconMenuItem(ICON_LC_UNDO, _("Undo"), "undo", mementoHandler.canUndo(), &listener);
 			ImGui::CommandIconMenuItem(ICON_LC_REDO, _("Redo"), "redo", mementoHandler.canRedo(), &listener);
 			ImGui::Separator();
-			ImGui::CommandIconMenuItem(ICON_LC_COPY, _("Copy"), "copy", hasSelection, &listener);
+			ImGui::CommandIconMenuItem(ICON_LC_CLIPBOARD_COPY, _("Copy"), "copy", hasSelection, &listener);
 			ImGui::CommandIconMenuItem(ICON_LC_SCISSORS, _("Cut"), "cut", hasSelection, &listener);
 			ImGui::CommandIconMenuItem(ICON_LC_CLIPBOARD_PASTE, _("Paste at reference"), "paste",
 									   _sceneMgr->clipboardData(), &listener);